requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.15
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.2
langchain-nvidia-ai-endpoints>=0.0.3

# RAG specific
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf>=3.17.1
unstructured>=0.10.30

# Web Search specific
duckduckgo-search>=4.1.1
tavily-python>=0.2.5
google-api-python-client>=2.108.0

# Local LLM support
ollama>=0.1.6

# Existing project dependencies
playwright>=1.40.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2

# --- Added for RAG Learning System ---
chromadb>=0.4.22 # Vector Database for interaction logs

# Add commentjson for robust JSON parsing
commentjson

# Add pysqlite3-binary to provide a newer sqlite3 version for ChromaDB
pysqlite3-binary

# Travel APIs
amadeus